Prefetch game images
This fixes an issue where when revealing cells for the first time,
the field isn't re-rendered all at once. For example, before this
change, on the first re-render, only the 1s and 2s would be revealed,
and then on the second re-render all of the 0s are revealed. It's very
subtle, but definitely noticeable if the requests take a bit too long.
This way, all of the cell images are preloaded before the first click,
and the new state renders all at once.